+++ src/Arch/Arc/Registers.cs
@@ -79,10 +79,10 @@ namespace Reko.Arch.Arc
             LpEnd = sysFactory.Reg("LP_END", PrimitiveType.Word32);
             AuxMacmode = sysFactory.Reg("AUX_MACMODE", PrimitiveType.Word32);
 
-            Z = new FlagGroupStorage(Status32, (uint) FlagM.ZF, "Z");
-            N = new FlagGroupStorage(Status32, (uint) FlagM.NF, "N");
-            C = new FlagGroupStorage(Status32, (uint) FlagM.CF, "C");
-            V = new FlagGroupStorage(Status32, (uint) FlagM.VF, "V");
+            Z = new FlagGroupStorage(Status32, (ulong) FlagM.ZF, "Z");
+            N = new FlagGroupStorage(Status32, (ulong) FlagM.NF, "N");
+            C = new FlagGroupStorage(Status32, (ulong) FlagM.CF, "C");
+            V = new FlagGroupStorage(Status32, (ulong) FlagM.VF, "V");
             S = new FlagGroupStorage(AuxMacmode, (uint) AuxFlagM.Sat, "S");
             ZNCV = new FlagGroupStorage(Status32, (uint) (FlagM.ZF | FlagM.NF | FlagM.CF | FlagM.VF), "ZNCV");
             ZNC = new FlagGroupStorage(Status32, (uint) (FlagM.ZF | FlagM.NF | FlagM.CF), "ZNC");
