+++ src/Arch/Arm/AArch64/Registers.cs
@@ -195,14 +195,14 @@ namespace Reko.Arch.Arm.AArch64
                     new [] { fpsr },
                 }).ToArray();
 
-            C = new FlagGroupStorage(pstate, (uint) FlagM.CF, nameof(C));
-            N = new FlagGroupStorage(pstate, (uint) FlagM.NF, nameof(N));
+            C = new FlagGroupStorage(pstate, (ulong) FlagM.CF, nameof(C));
+            N = new FlagGroupStorage(pstate, (ulong) FlagM.NF, nameof(N));
             NV = new FlagGroupStorage(pstate, (uint) (FlagM.NF|FlagM.VF), nameof(NV));
             NZ = new FlagGroupStorage(pstate, (uint) (FlagM.NF|FlagM.ZF), nameof(NZ));
             NZV = new FlagGroupStorage(pstate, (uint) (FlagM.NF|FlagM.ZF|FlagM.VF), nameof(NZV));
             NZCV = new FlagGroupStorage(pstate, (uint) (FlagM.NF|FlagM.ZF|FlagM.CF|FlagM.VF), nameof(NZCV));
-            V = new FlagGroupStorage(pstate, (uint) FlagM.VF, nameof(V));
-            Z = new FlagGroupStorage(pstate, (uint) FlagM.ZF, nameof(Z));
+            V = new FlagGroupStorage(pstate, (ulong) FlagM.VF, nameof(V));
+            Z = new FlagGroupStorage(pstate, (ulong) FlagM.ZF, nameof(Z));
             ZC = new FlagGroupStorage(pstate, (uint) (FlagM.ZF|FlagM.CF), nameof(ZC));
         }
     }
