+++ src/Arch/Cray/Cray1Architecture.cs
@@ -68,7 +68,7 @@ namespace Reko.Arch.Cray
             throw new NotImplementedException();
         }
 
-        public override FlagGroupStorage GetFlagGroup(RegisterStorage flagRegister, uint grf)
+        public override FlagGroupStorage GetFlagGroup(RegisterStorage flagRegister, ulong grf)
         {
             throw new NotImplementedException();
         }
@@ -93,7 +93,7 @@ namespace Reko.Arch.Cray
             throw new NotImplementedException();
         }
 
-        public override string GrfToString(RegisterStorage flagregister, string prefix, uint grf)
+        public override string GrfToString(RegisterStorage flagregister, string prefix, ulong grf)
         {
             throw new NotImplementedException();
         }
